+++ packages/core/src/__tests__/Router/edge-cases.test.js
