+++ packages/core/src/__tests__/Router/naming-routes.test.js
